backend/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=8.0.0
flask>=2.0.0
flask-cors>=3.0.0
waitress>=2.0.0
requests==2.31.0